requirements.txt
autogen-core==0.4.7
autogen-agentchat==0.4.7
autogen-ext[azure]==0.4.7
autogen-ext[openai]==0.4.7
autogen-ext[docker]==0.4.7
python-dotenv
pandas
matplotlib
rich